feat(strata-crypto): add musig2 sig creation
